+++ dune/common/std/make_array.hh
@@ -8,6 +8,9 @@
 #include <experimental/array>
 #endif
 
+#warning make_array.hh is deprecated and will be removed after Dune 2.8. \
+  Use deduction guide of `std::array` or `std::to_array`.
+
 namespace Dune {
 namespace Std {